Thread safety (#18)
Added thread safety which guarantees that the singleton and transient
configurations resolve properly in concurrent programs. That is,
singleton configurations should always result in one object and
transient should always create new objects.

Removed 'resolving' flag which was used to detect recursive
dependencies. The reason is that thread-safety requires locking
mechanisms because of which the resolving flag became useless.